Updated so that Staff can only edit while admin can edit and delete users.

+++ layer_presentation/userList.php
@@ -46,12 +46,19 @@ $users = $controller->listUsers();
                             <td><?php echo htmlspecialchars($user['full_name']); ?></td>
                             <td><?php echo htmlspecialchars($user['role']); ?></td>
                             <td>
-                                <a href="layer_presentation/editUser.php?id=<?php echo $user['id']; ?>" class="btn btn-warning btn-sm">
+                                
+                                <?php if (isset($_SESSION['role']) && $_SESSION['role'] == 'staff'): ?>
+                                    <a href="layer_presentation/editUser.php?id=<?php echo $user['id']; ?>" class="btn btn-warning btn-sm">
                                     <i class="fas fa-edit"></i>
-                                </a>
-                                <a href="layer_presentation/deleteUser.php?id=<?php echo $user['id']; ?>" class="btn btn-danger btn-sm">
-                                    <i class="fas fa-trash"></i>
-                                </a>
+                                    </a>
+
+                                    <?php if (isset($_SESSION['role']) && $_SESSION['role'] == 'admin'): ?>
+                                        <a href="layer_presentation/deleteUser.php?id=<?php echo $user['id']; ?>" class="btn btn-danger btn-sm">
+                                        <i class="fas fa-trash"></i>
+                                        </a>
+                                    <?php endif; ?>
+                                <?php endif; ?>
+
                             </td>
                         </tr>
                     <?php endforeach; ?>
